Generated by Spark: There should only be a single list of compliance results, not split between languages. There should only be one validation list for any language no need to pick between basic, javascript, python. All of the results should be listed on one page.
Print the description and rate the repo description great, good, poor, missing

for LICENSE only print the copyright holder and the short name of what you think the license is.

Add a hyperlink so we can go directly to the files that are found too for LICENSE, Contributing, Code of conduct, etc.

For dependency analysis if there are any copyleft licenses those need a warning required review